backend/requirements.txt
fastapi[all]
pymongo
python-dotenv
cryptography
bcrypt
pyjwt